+++ test/module-test.js
@@ -410,6 +410,63 @@ vows.describe('module tests').addBatch({
       assert.instanceOf(minified.sourceMap, SourceMapGenerator);
     }
   },
+  'arbitrary property transformations': {
+    'allows changing property value': {
+      'topic': function () {
+        return new CleanCSS({
+          level: {
+            1: {
+              transform: function (propertyName, propertyValue) {
+                if (propertyName == 'background-image' && propertyValue.indexOf('/path/to') > -1) {
+                  return propertyValue.replace('/path/to', '../valid/path/to');
+                }
+              }
+            }
+          }
+        }).minify('.block{background-image:url(/path/to/image.png);border-image:url(image.png)}');
+      },
+      'gives right output': function (error, output) {
+        assert.equal(output.styles, '.block{background-image:url(../valid/path/to/image.png);border-image:url(image.png)}');
+      }
+    },
+    'allows dropping properties': {
+      'topic': function () {
+        return new CleanCSS({
+          level: {
+            1: {
+              transform: function (propertyName) {
+                if (propertyName.indexOf('-o-') === 0) {
+                  return false;
+                }
+              }
+            }
+          }
+        }).minify('.block{-o-border-radius:2px;border-image:url(image.png)}');
+      },
+      'gives right output': function (error, output) {
+        assert.equal(output.styles, '.block{border-image:url(image.png)}');
+      }
+    },
+    'combined with level 2 optimization': {
+      'topic': function () {
+        return new CleanCSS({
+          level: {
+            1: {
+              transform: function (propertyName) {
+                if (propertyName == 'margin-bottom') {
+                  return false;
+                }
+              }
+            },
+            2: true
+          }
+        }).minify('.block{-o-border-radius:2px;margin:0 12px;margin-bottom:5px}');
+      },
+      'drops property before level 2 optimizations': function (error, output) {
+        assert.equal(output.styles, '.block{-o-border-radius:2px;margin:0 12px}');
+      }
+    }
+  },
   'accepts a list of source files as array': {
     'relative': {
       'with rebase to the current directory': {